code refactoring for riichi and rng
