+++ src/core/mahjong/rng.hpp
@@ -1,9 +1,27 @@
 
 #include "core/adt/itertools.hpp"
 #include "core/adt/vector.hpp"
-#include "game.hpp"
+#include "hand.hpp"
 
 namespace mj {
+class TileSrc
+{
+public:
+    virtual ~TileSrc() = default;
+
+    /**
+     * @return Tile draw the next tile for the player at direction dir.
+     */
+    virtual Tile operator()(Dir dir) = 0;
+
+    /**
+     * @return Tile draw a tile without changing the number of live tiles
+     * remaining.
+     */
+    virtual Tile dora() = 0;
+};
+
+
 namespace random {
 using WallContainer = vector<Tile, k_DeckSize>;
 using WallContainer34 = vector<int, k_DeckSize>;
