+++ test/sims/riichi.cc
@@ -1,0 +1,17 @@
+
+#include "test/test.hpp"
+#include "sims/riichi.hpp"
+
+TEST t_RiichiBasic()
+{
+    // set seed 
+    std::mt19937 _rng{10};
+    mj::random::Rng34 _deck(_rng);
+    // find the first tile that's 3s
+};
+
+int main()
+{
+
+    return g_FailureCount;
+}
